requirements.txt
anyio
APScheduler
beautifulsoup4
blinker
bs4
cachetools
certifi
charset-normalizer
click
colorama
Flask
Flask-APScheduler
gunicorn
h11
httpcore
httpx
idna
itsdangerous
Jinja2
lxml
MarkupSafe
python-dateutil
python-telegram-bot
pytz
pytz-deprecation-shim
requests
six
sniffio
soupsieve
tornado
tzdata
tzlocal
urllib3
Werkzeug
curl_cffi
cloudscraper
